requirements.txt
resampy
python_speech_features
scikit-learn
tensorflow-gpu>=1.15.2
